+++ changelog.md
@@ -9,6 +9,7 @@ v1.16.0:
 - Fix 90-degree rotated bar wiggling in 12.0+. Note that this is still a largely unsupported feature and requires a lot of manual manipulation of ancillary module items (icons, text labels, etc.), but the core "the bar wiggles like crazy" problem is now gone.
 - Fix bar texture overrides not always applying immediately after enabling the option (previously you sometimes had to change the texture before the override would take effect).
 - Update TBC Anniversary TOC.
+- Fix "locked" alpha text showing even when the bar was at 0% alpha (via the Transparency Settings cases in the module's options).
 
 v1.15.23: